Factor out groups API functional tests into multiple modules
